Adds the ability to test Symfony commands
